[WIP] [java] InnerHandler と objectConstructor の用法: mockk,
